requirements.txt
cachetools==5.5.0
Flask==3.0.3
psycopg2-binary==2.9.9
pytest==8.3.3
python-dotenv==1.0.1
python-telegram-bot==21.6
requests>=2.32.3
SQLAlchemy==2.0.35
flasgger==0.9.7.1